+++ templates/admin/base_admin.html.twig
@@ -53,7 +53,7 @@
 								</a>
 
 
-								<a href="{{path('dashbaord_admin_list_users')}}" class="flex items-center px-2 py-2 text-base font-medium text-indigo-100 rounded-md hover:text-white hover:bg-indigo-600 group" x-state-description="undefined: &quot;bg-indigo-800 text-white&quot;, undefined: &quot;text-indigo-100 hover:text-white hover:bg-indigo-600&quot;">
+								<a href="{{path('users_admin_list')}}" class="flex items-center px-2 py-2 text-base font-medium text-indigo-100 rounded-md hover:text-white hover:bg-indigo-600 group" x-state-description="undefined: &quot;bg-indigo-800 text-white&quot;, undefined: &quot;text-indigo-100 hover:text-white hover:bg-indigo-600&quot;">
 									<svg class="w-6 h-6 mr-4 text-indigo-200" x-description="Heroicon name: outline/user-group" xmlns="http://www.w3.org/2000/svg" fill="none" viewbox="0 0 24 24" stroke="currentColor" aria-hidden="true">
 										<path stroke-linecap="round" stroke-linejoin="round" stroke-width="2" d="M17 20h5v-2a3 3 0 00-5.356-1.857M17 20H7m10 0v-2c0-.656-.126-1.283-.356-1.857M7 20H2v-2a3 3 0 015.356-1.857M7 20v-2c0-.656.126-1.283.356-1.857m0 0a5.002 5.002 0 019.288 0M15 7a3 3 0 11-6 0 3 3 0 016 0zm6 3a2 2 0 11-4 0 2 2 0 014 0zM7 10a2 2 0 11-4 0 2 2 0 014 0z"></path>
 									</svg>
@@ -93,7 +93,7 @@
 									Home
 								</a>
 
-								<a href="{{path('dashbaord_admin_list_users')}}" class="flex items-center px-2 py-2 text-sm font-medium leading-6 text-indigo-100 rounded-md hover:text-white hover:bg-indigo-600 group" x-state-description="undefined: &quot;bg-indigo-800 text-white&quot;, undefined: &quot;text-indigo-100 hover:text-white hover:bg-indigo-600&quot;">
+								<a href="{{path('users_admin_list')}}" class="flex items-center px-2 py-2 text-sm font-medium leading-6 text-indigo-100 rounded-md hover:text-white hover:bg-indigo-600 group" x-state-description="undefined: &quot;bg-indigo-800 text-white&quot;, undefined: &quot;text-indigo-100 hover:text-white hover:bg-indigo-600&quot;">
 									<svg class="w-6 h-6 mr-4 text-indigo-200" x-description="Heroicon name: outline/user-group" xmlns="http://www.w3.org/2000/svg" fill="none" viewbox="0 0 24 24" stroke="currentColor" aria-hidden="true">
 										<path stroke-linecap="round" stroke-linejoin="round" stroke-width="2" d="M17 20h5v-2a3 3 0 00-5.356-1.857M17 20H7m10 0v-2c0-.656-.126-1.283-.356-1.857M7 20H2v-2a3 3 0 015.356-1.857M7 20v-2c0-.656.126-1.283.356-1.857m0 0a5.002 5.002 0 019.288 0M15 7a3 3 0 11-6 0 3 3 0 016 0zm6 3a2 2 0 11-4 0 2 2 0 014 0zM7 10a2 2 0 11-4 0 2 2 0 014 0z"></path>
 									</svg>
