+++ templates/admin/index.html.twig
@@ -66,7 +66,7 @@
 					</div>
 					<div class="px-5 py-3 bg-gray-50">
 						<div class="text-sm">
-							<a href="{{path('dashbaord_admin_list_users')}}" class="font-medium text-cyan-700 hover:text-cyan-900">
+							<a href="{{path('users_admin_list')}}" class="font-medium text-cyan-700 hover:text-cyan-900">
 								View all
 							</a>
 						</div>
